Initial UI changes and API creation. Throwing error on db search
