+++ app/api/events/get-information/route.ts
@@ -1,0 +1,8 @@
+import { fetchEventById } from "@/app/lib/data";
+import { NextResponse } from "next/server";
+
+export async function POST(req: Request) {
+	const id = await req.json();
+	const event = await fetchEventById(id);
+	return NextResponse.json(event)
+}
