--- app/components/events-page/intermediate-layer.tsx
@@ -1,12 +1,0 @@
-// "use server"
-
-// import EventModalWithOnClose from "./event-modal-with-onClose";
-// import { fetchEventsById } from "@/app/lib/data";
-
-// export default async function IntermediateLayer({id}: {id: string}) {
-//   const event = await fetchEventsById(id)
-//   return (
-//     <div></div>
-//     // <EventModalWithOnClose event={event}/>
-//   )
-// }
