+++ app/components/skeletons/event-info-page.tsx
@@ -1,6 +1,6 @@
-export default function CardSkeleton() {
+export default function EventInfoPageSkeleton() {
 	return (
-		<div className="bg-transparent p-6 rounded-lg flex flex-row items-center justify-between border-2 border-blue-100 min-w-[800px]">
+		<div className="bg-transparent p-6 rounded-lg flex flex-row items-center justify-between border-2 border-blue-100">
 			{/* Left Side: Partner Info */}
 			<div className="flex flex-col space-y-4 w-2/3 pr-4">
 				<h2 className="h-6 w-3/4 bg-glowing-spot bg-no-repeat bg-[length:200%_100%] animate-glow rounded bg-gray-500"></h2>
