+++ app/lib/data.ts
@@ -77,18 +77,21 @@ export async function fetchUserEvents(organiser_uid: string) {
 }
 
 
-// export async function fetchEventsById(id: string) {
-// 	try {
-// 		const data = await sql<SQLEvent>`
-// 			SELECT * FROM events
-// 			WHERE id = ${id}
-// 		`;
-// 		return convertSQLEventToEvent(data.rows[0]);
-// 	} catch (error) {
-// 		console.error('Database error:', error);
-// 		throw new Error('Failed to fetch upcoming events');
-// 	}
-// }
+export async function fetchEventById(id: string) {
+	try {
+		const data = await sql<SQLEvent>`
+			SELECT *
+			FROM events
+			WHERE id::text LIKE '%' || ${id};
+		`;
+		console.log(data.rows)
+		return data
+		// return convertSQLEventToEvent(data.rows[0]);
+	} catch (error) {
+		console.error('Database error:', error);
+		throw new Error('Failed to fetch event');
+	}
+}
 
 export async function insertEvent(event: SQLEvent) {
 	try {
